docs(README-developers): Add workflow for the branch management
The documentation is a proposed method to handle development in the
master branch. Each pull request is automatically deployed to the
testing namespace. The testing images are manually deployed to the
production namespace.

It also proposes a workflow where each developer manages their own fork
with personal pull request branches. The new feature branches can then
be merged to the master.

Signed-off-by: Pauli <suokkos@gmail.com>